move profile image deletion signal to signals.py and register via AppConfig
