+++ Backend/accounts/models.py
@@ -5,10 +5,7 @@ from datetime import timedelta
 from PIL import Image
 import os
 import logging
-from django.db.models.signals import post_delete
-from django.dispatch import receiver
-from pathlib import Path
-from django.conf import settings
+
 
 logger = logging.getLogger(__name__) 
 
@@ -106,15 +103,4 @@ class OTP(models.Model):
     def is_valid(self):
         return timezone.now() < self.expires_at
 
-@receiver(post_delete, sender=CustomUser)
-def delete_profile_image_on_user_delete(sender, instance, **kwargs):
-    """Delete profile image file from disk when user is deleted securely."""
-    if instance.profile_image and instance.profile_image.path:
-        try:
-            file_path = Path(instance.profile_image.path).resolve()
-            media_root = Path(settings.MEDIA_ROOT).resolve()
 
-            if media_root in file_path.parents and file_path.is_file():
-                file_path.unlink()
-        except Exception as e:
-            logger.warning(f"Failed to securely delete profile image for user '{instance.email}': {e}")
